Vault key requirement
Vault uses same key color system as cardreaders for doors. Moved the relevant Enum to it's own class so it can be reached directly instead of through CardReader.cs. 
Key becomes required after choosing the Keycard Color from the list of Vault (script). The (small) color indicator on the vault is automatic, the object doesn't have to be dragged into the vault (just don't change the name).